use different redis connections for write/read operations to avoid landing all operation on the primary instance (#42)
use different redis connections for write/read operations to avoid landing all operation on the primary instance (#42)

# Issue Description

We noticed that most of our connections to a Redis cluster via cachette end up in the primary instance, instead of the replicas.

When reading or writing data to Redis we do use the same Redis connection, which is initialized in the `WriteThroughCache` constructor. 

If this connection is established to a replica and we try to do a write operation through the `WriteThroughCache` class then we will receive a READONLY Error.

if we receive a READONLY error we will keep trying to connect to the cluster until we have a connection with the primary instance (find more details here https://www.npmjs.com/package/ioredis#reconnect-on-error)

```js
    this.redisClient = new Redis(redisUrl, {
      retryStrategy: () => RedisCache.RETRY_DELAY,
      // master failover
      reconnectOnError: (err: any) => err.message.startsWith('READONLY'),
      // This will prevent the get/setValue calls from hanging
      // if there is no active connection.
      enableOfflineQueue: false,
    });
```

# Suggested Solution

Use two different Redis connections, one of them will be used only for write/update/delete operations and the 2nd connection will be used for the read operations only.