💨 Support plain object as key for caching (#56)
When we have a plain object as an argument in a function call that is cached, we are not able to create different keys as cacheKey even if the object is not the same in 2 different calls. This behavior results in fetching the wrong result for a call that has been cached before.

This PR introduces the support of plain object in buildCacheKey function.

We basically get all the values of the plain object and join them with - char. In the case of:

getFoo(
	arg1 = 'coucou',
	arg2 = { nestedProp: 'nestedValue', nestedProp2: 'nestedValue2' },
)
Will lead to this cacheKey: getFoo-coucou-nestedProp-nestedValue-nestedProp2-nestedValue2.

This PR include the support for undefined, null and array as well!