🔧 Re-emit Redlock clientErrors (#59)
While reviewing Redlock's error handling, we noticed it doesn't throw/reject
much. Instead, it emits a `clientError`, but cachette does nothing with it,
leaving `cachette` users (including us) in the dark.

This PR simply re-emits these `clientErrors` as a `warn` event, letting
cachette users listen to it and do something with it (at least: log).

By the way, reviewing Redlock' error handling got us to notice that their
error handling is going to change **A LOT** in the upcoming Redlock v5.
So, adding a little gotcha test to shout at our future selves when doing
this upgrade.

By the way by the way, relock.