💨 Use the new way of importing node's EventEmitter (#73)
Node changed the name of the library from `events` to `node:events` with
v16. It still has backward compatibilities it seems, but importing the
cachette library failed on one occasion: the result was that the
CacheInstance was not resolved as an EventEmitter.

The `package.json` node version is already at `>=18`, so this should be
fine.