💨 Expose wait command (#83)
Exposing Redis' WAIT command which can be useful to ensure data
freshness across all nodes of a cluster.

Adding arbitrary default replica of 5 (the maximum supported by AWS
Elasticache) and timeout to 200 millisecond. These could be parametrized
from the client. Recommending to keep the timeout as low as possible
since this is a blocking operation. The client can know whether or not
the expected number of replicas acknowledged the last command by
comparing the return value of `waitForReplication`. It will return the
number of replicas that acknowledge

---------

Co-authored-by: AdelUnito <adelb@unito.io>
Co-authored-by: Ronan Jouchet <ronanj@unito.io>